+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix: les API Start et Updatestatus ne doivent pas retourner de donnée personnelle.
 * L'api Updatestatus retourne le statut de la transaction.
 * Nouvelle API `convert`.
+* Divers fix.
 
 ## 0.0.4